Refactored ScalaFX constants in separate file
