+++ src/main/scala/view/scalaFX/FXControllers/MutationsPanelController.scala
@@ -5,6 +5,7 @@ import model.genome.Genes._
 import model.mutation.Mutation
 import scalafx.scene.control.{Button, Label}
 import scalafxml.core.macros.sfxml
+import view.scalaFX.ScalaFXConstants.Style.MutationsChoice.{CHOSEN_BUTTON_STYLE, OTHER_BUTTON_STYLE}
 
 sealed trait MutationsPanelControllerInterface {
 
@@ -29,9 +30,6 @@ class MutationsPanelController(
     private val furColorDominantChoiceButton: Button,
     private val mutationIncomingLabel: Label
 ) extends MutationsPanelControllerInterface {
-
-  val CHOSEN_STYLE: String = "chosen-button"
-  val OTHER_STYLE: String = "dashed-button"
 
   def furColorDominantChoiceClick(): Unit = manageChoiceClick(
     FUR_COLOR,
@@ -133,8 +131,8 @@ class MutationsPanelController(
   }
 
   private def updateButtonStyle(clicked: Button, other: Button): Unit = {
-    clicked.styleClass += CHOSEN_STYLE
-    other.styleClass += OTHER_STYLE
+    clicked.styleClass += CHOSEN_BUTTON_STYLE
+    other.styleClass += OTHER_BUTTON_STYLE
   }
 
   private def showMutationIncoming(): Unit = mutationIncomingLabel.visible = true
@@ -152,8 +150,8 @@ class MutationsPanelController(
 
   private def resetButtons(btns:Seq[Button]): Unit = {
     btns.foreach(btn => {
-      btn.styleClass -= CHOSEN_STYLE
-      btn.styleClass -= OTHER_STYLE
+      btn.styleClass -= CHOSEN_BUTTON_STYLE
+      btn.styleClass -= OTHER_BUTTON_STYLE
       btn.disable = false
     })
   }
