+++ src/main/scala/view/scalaFX/components/charts/pedigree/BunnyPedigreeView.scala
@@ -5,7 +5,7 @@ import scalafx.geometry.Insets
 import scalafx.scene.image.{Image, ImageView}
 import scalafx.scene.layout._
 import scalafx.scene.text.Text
-import view.scalaFX.ScalaFxViewConstants.GenealogicalTree.{BUNNY_ALLELE_PADDING, FONT_INFO_PERCENT, TREE_INFO_PROPORTION}
+import view.scalaFX.ScalaFXConstants.GenealogicalTree.{BUNNY_ALLELE_PADDING, FONT_INFO_PERCENT, TREE_INFO_PROPORTION}
 import view.scalaFX.components.charts.pedigree.PedigreeChart.{bunnyIconSize, spacingRegion}
 import view.scalaFX.utilities.Direction.Right
 import view.scalaFX.utilities.{BunnyImageUtils, Direction, ImageType}
